requirements.txt
keras
tensorflow-gpu
#tensorflow

matplotlib
scipy
scikit-learn
scikit-image
more-itertools
imutils

requests
graphviz
pydot
h5py
